@@ -1,11 +1,15 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { Provider } from 'react-redux';
 import App from './components/App';
 import UserProvider from './providers/UserProvider';
+import store from './redux/store';
 
 ReactDOM.render(
   <UserProvider>
-    <App />
+    <Provider store={store}>
+      <App />
+    </Provider>
   </UserProvider>,
   document.getElementById('root')
 );
